+++ src/colorpicker/multicolorpicker.cpp
@@ -96,7 +96,7 @@ void MultiColorPicker::updateColorStates(const std::vector<QColor>& colorSchemes
         std::vector<QColor> newScheme(mMaxCount);
         // in cases where an light is currently showing one color but can show more, set all the
         // additional colors it can show as the one color
-        if (colorSchemes.size() < mCount) {
+        if (colorSchemes.size() < mMaxCount) {
             for (std::size_t i = 0u; i < colorSchemes.size(); ++i) {
                 newScheme[i] = colorSchemes[i];
             }
@@ -104,9 +104,10 @@ void MultiColorPicker::updateColorStates(const std::vector<QColor>& colorSchemes
                 newScheme[i] = colorSchemes[0];
             }
         } else {
-            newScheme = colorSchemes;
+            for (std::size_t i = 0u; i < mMaxCount; ++i) {
+                newScheme[i] = colorSchemes[i];
+            }
         }
-
         mScheme = newScheme;
         mColorSchemeCircles->updateColorScheme(newScheme, true);
     }